+++ templates/controllers/controller-template.ts
@@ -15,6 +15,10 @@ export async function create(req: Request, res: Response, next: NextFunction) {
 export async function deleteReplace_Here(req: Request, res: Response, next: NextFunction) {
     try {
         const { id } = req.params;
+        const oldData = await replace_hereRepository.get(id);
+        if(!oldData){
+            return res.status(404).send({ message: 'Not found.' });
+        }
         const response = await replace_hereRepository.deleteReplace_Here(id);
         return res.status(202).send(response);
     } catch (error) {
@@ -27,6 +31,9 @@ export async function get(req: Request, res: Response, next: NextFunction) {
         const userData: UserData = req.cookies.userData;
         const id = req.params?.id || userData.id;
         const response = await replace_hereRepository.get(id);
+        if(!response){
+            return res.status(404).send({ message: 'Not found.' });
+        }
         return res.status(200).json(response);
     } catch (error) {
         return next(error);
@@ -38,6 +45,11 @@ export async function update(req: Request, res: Response, next: NextFunction) {
         const userData: UserData = req.cookies.userData;
         const data = req.body;
         const id = req.params?.id || userData.id;
+
+        const oldData = await replace_hereRepository.get(id);
+        if(!oldData){
+            return res.status(404).send({ message: 'Not found.' });
+        }
 
         const response = await replace_hereRepository.update(id, data);
         return res.status(200).json(response);
